# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/projects/{id}": {
      "get": {
        "description": "Retrieves a project by its ID. Users can only access their own projects.",
        "operationId": "ProjectController_getProjectById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Access denied to this project"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get project by ID",
        "tags": [
          "projects"
        ]
      },
      "put": {
        "description": "Updates an existing project. Users can only update their own projects.",
        "operationId": "ProjectController_updateProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Project update data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Access denied to this project"
          },
          "404": {
            "description": "Project not found"
          },
          "409": {
            "description": "Project with this unique ID already exists"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update project",
        "tags": [
          "projects"
        ]
      },
      "delete": {
        "description": "Deletes a project. Users can only delete their own projects.",
        "operationId": "ProjectController_deleteProject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Access denied to this project"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/chat/project/{projectId}": {
      "get": {
        "description": "Retrieves a paginated list of all chats in a specific project",
        "operationId": "ChatController_getProjectChats",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "adminId",
            "required": false,
            "in": "query",
            "description": "Filter chats by admin ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": false,
            "in": "query",
            "description": "Filter chats by participant ID",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of chats to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved project chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all chats for a specific project",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/chat/project/{projectId}/participant/{participantId}": {
      "post": {
        "description": "Creates a new chat or returns existing one between admin and participant",
        "operationId": "ChatController_createOrGetChat",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "ID of the participant",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved or created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or get chat with participant",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/project-participants": {
      "post": {
        "description": "Creates a new participant for a specific project. Admin authentication required.",
        "operationId": "ProjectParticipantController_createParticipant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectParticipantDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created participant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectParticipantResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid participant data or participant ID already exists"
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new project participant (Admin)",
        "tags": [
          "project-participants"
        ]
      }
    },
    "/api/project-participants/project/{projectId}": {
      "get": {
        "description": "Retrieves a paginated list of all participants for a specific project",
        "operationId": "ProjectParticipantController_getProjectParticipants",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of participants to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of participants to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved participants",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectParticipantResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all participants for a project",
        "tags": [
          "project-participants"
        ]
      }
    },
    "/api/project-participants/{participantId}": {
      "get": {
        "description": "Retrieves information about a specific project participant",
        "operationId": "ProjectParticipantController_getParticipantById",
        "parameters": [
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "ID of the participant",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved participant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectParticipantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "404": {
            "description": "Participant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get participant by ID",
        "tags": [
          "project-participants"
        ]
      },
      "delete": {
        "description": "Removes a participant from a project",
        "operationId": "ProjectParticipantController_deleteParticipant",
        "parameters": [
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "ID of the participant to delete",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted participant"
          },
          "401": {
            "description": "User is not authenticated"
          },
          "404": {
            "description": "Participant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete project participant",
        "tags": [
          "project-participants"
        ]
      }
    },
    "/api/project-participants/project/{projectId}/unique/{uniqueId}": {
      "get": {
        "description": "Retrieves information about a participant using their unique identifier within a project",
        "operationId": "ProjectParticipantController_getParticipantByUniqueId",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "uniqueId",
            "required": true,
            "in": "path",
            "description": "Unique identifier of the participant",
            "schema": {
              "example": "user123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved participant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectParticipantResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "404": {
            "description": "Participant not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get participant by unique ID",
        "tags": [
          "project-participants"
        ]
      }
    },
    "/api/project-participants/project/{projectId}/count": {
      "get": {
        "description": "Returns the total number of participants in a project",
        "operationId": "ProjectParticipantController_getParticipantsCount",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved participants count",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "example": 15
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get participants count",
        "tags": [
          "project-participants"
        ]
      }
    },
...
    "/api/mobile-chat/project/{projectId}/participant/{participantId}": {
      "post": {
        "description": "Gets or creates a chat between admin and participant. No authentication required for mobile app usage.",
        "operationId": "MobileChatController_getOrCreateChat",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "External participant ID (string) provided by mobile app",
            "schema": {
              "example": "mobile_user_123",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved or created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid project or participant data"
          }
        },
        "summary": "Get or create chat (Mobile App)",
        "tags": [
          "mobile-chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat CRM API",
    "description": "API documentation for Chat CRM application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "Admin"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "User"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "admin@chat-crm.com"
          },
          "role": {
            "type": "string",
            "description": "User role (always Admin)",
            "enum": [
              "Admin",
              "Participant"
            ],
            "example": "Admin"
          },
          "createdAt": {
            "type": "string",
            "description": "User creation date",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "User last update date",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "email",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ProjectResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "My Awesome Project"
          },
          "uniqueId": {
            "type": "string",
            "description": "Unique project ID assigned by admin",
            "example": "PROJ-001"
          },
          "createdBy": {
            "description": "User who created the project",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserResponseDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "description": "Project creation date",
            "example": "2023-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "Project last update date",
            "example": "2023-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "uniqueId",
          "createdBy",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "Updated Project Name",
            "minLength": 1,
            "maxLength": 255
          },
          "uniqueId": {
            "type": "string",
            "description": "Unique project ID assigned by admin",
            "example": "PROJ-002",
            "minLength": 3,
            "maxLength": 50
          }
        }
      },
...
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Chat ID",
            "example": 1
          },
          "projectId": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "project": {
            "description": "Project information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectDto"
              }
            ]
          },
          "admin": {
            "description": "Admin user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "participant": {
            "description": "Participant user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat creation timestamp",
            "example": "2025-01-15T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat last update timestamp",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the chat is active",
            "example": true
          },
          "lastMessage": {
            "description": "Last message in the chat",
            "allOf": [
              {
                "$ref": "#/components/schemas/LastMessageDto"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "Number of unread messages for the requesting user",
            "example": 3
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last message",
            "example": "2025-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "project",
          "admin",
          "participant",
          "createdAt",
          "updatedAt",
          "isActive",
          "unreadCount"
        ]
      },
...
      "CreateProjectParticipantDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "number",
            "description": "ID of the project where the participant will be added",
            "example": 1,
            "minimum": 1
          },
          "participantId": {
            "type": "string",
            "description": "External participant ID provided by mobile app",
            "example": "mobile_user_123",
            "minLength": 3,
            "maxLength": 50
          },
          "firstName": {
            "type": "string",
            "description": "First name of the participant (optional for mobile app)",
            "example": "John",
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the participant (optional for mobile app)",
            "example": "Doe",
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "description": "Email address of the participant (optional for mobile app)",
            "example": "john.doe@example.com"
          }
        },
        "required": [
          "projectId",
          "participantId"
        ]
      },
      "ProjectParticipantResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Participant ID",
            "example": 1
          },
          "projectId": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "User ID",
            "example": 2
          },
          "participantId": {
            "type": "string",
            "description": "External participant ID provided by mobile app",
            "example": "mobile_user_123"
          },
          "firstName": {
            "type": "string",
            "description": "First name of the participant",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of the participant",
            "example": "Doe"
          },
          "email": {
            "type": "object",
            "description": "Email address of the participant (optional)",
            "example": "john.doe@example.com",
            "nullable": true
          },
          "isOnline": {
            "type": "boolean",
            "description": "Whether the participant is online",
            "example": false
          },
          "lastSeen": {
            "type": "object",
            "description": "Last seen timestamp",
            "example": "2025-01-15T10:00:00.000Z",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "When the participant was added to the project",
            "example": "2025-01-15T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "userId",
          "participantId",
          "firstName",
          "lastName",
          "email",
          "isOnline",
          "lastSeen",
          "createdAt"
        ]
      }
    }
  }
}